pyproject.toml
[project]
name = "deja-q"
version = "0.1.0"
description = "Deja Q is a Slack app that helps users find answers to their questions by automatically searching for similar questions that have been asked before"
authors = [
    {name = "Abhimanyu Yadav",email = "abhiy.iitk@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "slack-sdk (>=3.34.0,<4.0.0)",
    "python-dotenv (>=1.0.1,<2.0.0)",
    "flask (>=3.1.0,<4.0.0)",
    "slackeventsapi (>=3.0.3,<4.0.0)",
    "torch (==2.2.1)",
    "sentence-transformers (>=3.4.1,<4.0.0)",
    "numpy (>=1.24.0,<2.0.0)",
    "pytest (>=8.3.5,<9.0.0)",
    "pytest-mock (>=3.14.0,<4.0.0)"
]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
